Implement basic Workflow-Execution logic.
The function *executeStep* will first set the AtomicReference to the actual task which is executed.
After that, it will check if a task requires some additional parameter. If it needs some it will wait until they are provided.
After a task Execution, the function *postAction* retrieves the next element of the queue and the whole process starts again.