Implements a function to restart a workflow.
A second singleton is created which only contains the definitions of the workflows. As soon as a workflow is started or restarted, a deep copy of the object is created and returned. This prevents parameters from being set during cloning.
In addition, a new status for workflows has been implemented (Finished). A workflow is Finished if it no longer has any elements in the execution queue.