Switch the Http-Client from OkHttp3 to Spring Web Client.
Therefore we have to change the hole ERequestSender Class for using the Web Client. There are also a few new attributes added to the CRequest Class. This is necessary because with the WebClient Framework we have no access to the response header, so the response media type has to be extracted from the related RAML-file.
The WebClient now accepts only responses with the same Media-Type provided in the related RAML-file.