Implement the concept of multiple active workflows.
Now it is possible to start multiple workflows all based on the same model without the need of parsing the model for each new instance. All models are now stored in a separate enum. For starting a workflow the user must provide the name of the model and a unique name for identifying the specific instance.