Add a new workflow parser.
The workflow parser is now used to parse the model string from the modelling tool into a *WorkflowModel*. This is then added to the state and permanently saved. The *WorkflowModel* is then converted to an "executable" *CWorkflow* object by the *WorkflowConversionService*.

When the server starts all saved models are loaded in the state.